+++ CHANGELOG.md
@@ -6,6 +6,17 @@ The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.1.0/),
 and this project adheres to
 [Semantic Versioning](https://semver.org/spec/v2.0.0.html).
 
+## [0.2.0]
+
+### Changed
+
+- [store] use cross-platform env var checking
+- [store-deno-kv] & [store-deno-kv-fs] export `getKv()` function
+
+### Added
+
+- [store-no-op] a no-op implementation of the interface
+
 ## [0.1.0]
 
 - Initial release on JSR